have it graphing the time domain signal and the frequency domain signal at the same time
Was having troubles figuring out what to do with the signal in order to
get the responses I wanted (changing colors, pulsing, moving, etc) so
got it graphing instead. There's lots of unused code in the main.js and
the visualization.js (for the 3D stuff I was trying) but it will
eventually get used when we start adding more spheres that will be
doing neat stuff. Just start up a simplehttpserver in that folder to
run it and click on the "load sample mp3" label.